Add capability: search campaign by id

+++ features/merchant/seller_shop_flash_sale/src/main/java/com/tokopedia/shop/flashsale/domain/usecase/GetSellerCampaignDetailUseCase.kt
@@ -13,7 +13,7 @@ class GetSellerCampaignDetailUseCase @Inject constructor(
 ) : GraphqlUseCase<CampaignUiModel>(repository) {
 
     suspend fun execute(
-        campaignId: Int,
+        campaignId: Long,
     ): CampaignUiModel {
         return coroutineScope {
             val campaignList = getSellerCampaignListUseCase.execute(
